+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # mahoraga-cli
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - mahoraga-agent@0.3.2
+
 ## 0.4.1
 
 ### Patch Changes